tabsbar added in layout
