+++ frontend/src/components/layouts/applayout.tsx
@@ -2,9 +2,10 @@ import React, { FunctionComponent, useEffect } from 'react'
 import Header from './header'
 import Footer from './footer'
 import Sidebar from './sidebar'
-import { Outlet } from "react-router-dom";
+import { Outlet, useLocation } from "react-router-dom";
 import { useAppSelector } from '../../redux/hooks';
 import { selectIsShowingSidebar } from '../../redux/configSlice';
+import TabsBar from './tabsbar';
 
 type PageLayoutPropType = {
 
@@ -12,7 +13,11 @@ type PageLayoutPropType = {
 
 const AppLayout: FunctionComponent<PageLayoutPropType> = () => {
 
+    const location = useLocation()
+
     const isShowingSidebar: boolean = useAppSelector(selectIsShowingSidebar)
+
+    let showTabsBar = (location.pathname.startsWith("/db")) ? true : false
 
     return (
         <React.Fragment>
@@ -20,8 +25,11 @@ const AppLayout: FunctionComponent<PageLayoutPropType> = () => {
                 <Header />
                 <div className="appcontent">
                     {isShowingSidebar && <Sidebar />}
-                    <main id="mainContainer" className={`maincontainer`}>
-                        <Outlet />
+                    <main className="maincontainer">
+                        {showTabsBar && <TabsBar />}
+                        <div id="maincontent" className="maincontent">
+                            <Outlet />
+                        </div>
                     </main>
                 </div>
                 <Footer />
