+++ frontend/src/components/layouts/tabsbar.module.scss
@@ -1,0 +1,3 @@
+.tabs {
+    margin-bottom: 0px !important;
+}
