+++ frontend/src/styles/index.scss
@@ -19,6 +19,7 @@ $table-row-active-background-color: $active-color;
 
 // Tabs
 $tabs-toggle-link-active-background-color: $active-color;
+$tabs-boxed-link-active-background-color: transparent;
 
 // Card
 $card-shadow: none;
